Remove tags from the single content pages.
As long as the tags are totally unrelated to the content next to them, and also don't appear in the form of an easily understandable word cloud, it makes no sense to have them on every content page.
We still keep them for the list of project news and blog posts.