Update 0.4.13 Download URL to latest fixed version. (#88)
I'm not a fan of putting too much effort into servicing an alpha version of an operating system. But original 0.4.13 can't even connect to RAPPS anymore due to the certificate change, and 0.4.14 is still delayed.
Joachim just released a fixed version of 0.4.13 and this commit introduces a "reactos-download-version" shortcode next to "reactos-version". We can therefore change the download files to "0.4.13-release-14-g2494cfc" while still advertising it as "ReactOS 0.4.13".
Imagine it like a Windows ISO with latest patches applied.